median, laplacian and Guassian filter code
This is my use of OpenCV code where I applied the median, Laplacian, and Gaussian filters. This is currently the basic input and output of the code and need to work on optimizing how to handle user input. I can easily make the change to download the final image but this is currently in work so that will be done last